requirements.txt
allure-pytest~=2.13.2
black~=23.7.0
faker~=19.6.0
pytest~=7.4.0
pytest-xdist~=3.3.1
requests~=2.31.0
selenium~=4.10.0
webdriver-manager~=4.0.0